DROOLS-4958: Changes required during code review
