+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-kogito-client/src/main/java/org/drools/workbench/screens/scenariosimulation/kogito/client/dmn/DMNTypeFactory.java
@@ -20,7 +20,6 @@ import java.util.List;
 import java.util.Map;
 
 import jsinterop.base.Js;
-import org.drools.workbench.screens.scenariosimulation.kogito.client.dmn.feel.BuiltInType;
 import org.kie.workbench.common.dmn.webapp.kogito.marshaller.js.model.dmn12.JSITItemDefinition;
 
 /**
@@ -80,59 +79,5 @@ public class DMNTypeFactory {
                                  !fields.isEmpty(),
                                  fields,
                                  null);
-    }
-
-
-    public static class ClientDMNType {
-
-        private String namespace;
-        private String name;
-        private String id;
-        private boolean collection;
-        private boolean composite;
-        private Map<String, ClientDMNType> fields;
-        private BuiltInType feelType;
-
-        public ClientDMNType(String namespace, String name, String id, boolean isCollection, BuiltInType feelType) {
-            this.namespace = namespace;
-            this.name = name;
-            this.id = id;
-            this.collection = isCollection;
-            this.feelType = feelType;
-        }
-
-        public ClientDMNType(String namespace, String name, String id, boolean isCollection, boolean isComposite, Map<String, ClientDMNType> fields, BuiltInType feelType) {
-            this(namespace, name, id, isCollection, feelType);
-            this.fields = fields;
-            this.composite = isComposite;
-        }
-
-        public String getNamespace() {
-            return namespace;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public String getId() {
-            return id;
-        }
-
-        public boolean isCollection() {
-            return collection;
-        }
-
-        public boolean isComposite() {
-            return composite;
-        }
-
-        public Map<String, ClientDMNType> getFields() {
-            return fields;
-        }
-
-        public BuiltInType getFeelType() {
-            return feelType;
-        }
     }
 }
